+++ packages/contracts-bedrock/src/libraries/BatchInboxErrors.sol
@@ -5,4 +5,4 @@ pragma solidity 0.8.15;
 error OnlyElectionWinner(address caller, address winner);
 
 /// @notice Error when the batch's target block is not the current block.
-error InvalidTargetBlock(uint targetBlock, uint currentBlock);
+error InvalidTargetBlock(uint256 targetBlock, uint256 currentBlock);
